[FEAT] 음주 리포트 API 로직 완성 #30

+++ src/main/java/com/umc/puppymode2/domain/report/dto/DrinkReportResponseDTO.java
@@ -9,8 +9,8 @@ import lombok.*;
 @Builder
 public class DrinkReportResponseDTO {
     private Integer goal;
-    private Integer drinkRecordCount;
-    private Integer drinkDays;
-    private Integer achievementRate;
-    private Integer scoldedCount;
+    private Long drinkRecordCount;
+    private Long drinkDays;
+    private int achievementRate;
+    private int scoldedCount;
 }
